moved sections to their own components.
